# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MagicMirror² Remote Control API",
    "description": "The MMM-Remote-Control Module for MagicMirror² implements a RESTful(-ish) API to control the MagicMirror² using the existing functionality built-in to MMM-Remote-Control, as well as the notifications commands built into most modules. In addition, the API creates a basic framework which allows for each module to expand or customize their own API by a simple notification.",
    "contact": {
      "name": "MMM-Remote-Control",
      "url": "https://github.com/Jopyth/MMM-Remote-Control"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/Jopyth/MMM-Remote-Control/blob/master/LICENSE.md"
    },
    "version": "2.3.0+"
  },
...
  "paths": {
...
    "/api/saves": {
      "get": {
        "tags": ["API"],
        "summary": "Return backups of the config file",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "query": {
                    "data": "saves"
                  },
                  "data": [
                    "2020-11-14T20:41:52.589Z",
                    "2020-11-14T20:41:13.876Z",
                    "2020-11-09T20:38:21.783Z",
                    "2020-11-09T20:32:59.994Z"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/classes/{:value}": {
      "get": {
        "tags": ["Module Control", "API"],
        "summary": "Show or use a class",
        "parameters": [
          {
            "name": ":value",
            "in": "path",
            "required": true,
            "allowEmptyValue": true,
            "description": "The class name to use or empty to list all available classes",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "query": {
                    "data": "classes"
                  },
                  "data": {
                    "test": {
                      "toggle": "calendar"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid class name provided",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid value test provided in request. Use /api/classes to see actual values"
                }
              }
            }
          }
        }
      }
    },
    "/api/command/{:value}": {
      "get": {
        "tags": ["Mirror Control", "API"],
        "summary": "Execute command saved in config",
        "parameters": [
          {
            "name": ":value",
            "in": "path",
            "required": true,
            "description": "The name of the custom command to execute",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully executed command"
          },
          "400": {
            "description": "Bad Request - Command value is required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Value is needed"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - API key required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "tags": ["API"],
        "summary": "Return the current MagicMirror² configuration",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved configuration",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "query": {
                    "data": "config"
                  },
                  "data": {
                    "address": "0.0.0.0",
                    "electronOptions": {
                      "fullscreen": false
                    },
                    "ipWhitelist": [],
                    "modules": [
                      {
                        "module": "MMM-Remote-Control",
                        "position": "top_right",
                        "config": {
                          "showModuleApiMenu": true
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/config/edit": {
      "post": {
        "tags": ["Mirror Control", "API"],
        "summary": "Modify the current MagicMirror² configuration",
        "security": [
          {
            "apikey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated configuration",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "query": {
                    "data": "config"
                  },
                  "data": {
                    "address": "0.0.0.0",
                    "electronOptions": {
                      "fullscreen": false
                    },
                    "ipWhitelist": [],
                    "modules": [
                      {
                        "module": "MMM-Remote-Control",
                        "position": "top_right",
                        "config": {
                          "showModuleApiMenu": true
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters in request body",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid URL provided in request body"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - API key required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/minimize": {
      "get": {
        "tags": ["Mirror Control", "API"],
        "summary": "Minimize the browser instance (Electron)",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully minimized window",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - API key required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
    "/api/togglefullscreen": {
      "get": {
        "tags": ["Mirror Control", "API"],
        "summary": "Toggle fullscreen (Electron)",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully toggled fullscreen mode",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - API key required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
    "/api/devtools": {
      "get": {
        "tags": ["Mirror Control", "API"],
        "summary": "Toggle devtools (Electron)",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully toggled devtools",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - API key required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/save": {
      "get": {
        "tags": ["API"],
        "summary": "Saves the MMM-Remote-Control defaults to a file.",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully saved defaults to file",
            "content": {
              "application/json": {
                "example": null
              }
            }
          },
          "403": {
            "description": "Forbidden - API key required",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}